+++ app/src/main/kotlin/ac/mdiq/podcini/ui/fragment/OnlineFeedFragment.kt
@@ -365,7 +365,9 @@ class OnlineFeedFragment : Fragment() {
         val textColor = MaterialTheme.colorScheme.onSurface
         val feedLogsMap_ = feedLogsMap!!
         Scaffold(topBar = { MyTopAppBar() }) { innerPadding ->
-            if (showProgress) CircularProgressIndicator(progress = { 0.6f }, strokeWidth = 10.dp, modifier = Modifier.padding(innerPadding).size(50.dp))
+            if (showProgress) Box(contentAlignment = Alignment.Center) {
+                CircularProgressIndicator(progress = {0.7f}, strokeWidth = 10.dp, color = textColor, modifier = Modifier.size(50.dp).align(Alignment.Center))
+            }
             else Column(modifier = Modifier.padding(innerPadding).fillMaxSize().padding(start = 10.dp, end = 10.dp)) {
                 if (showFeedDisplay) ConstraintLayout(modifier = Modifier.fillMaxWidth().height(100.dp).background(MaterialTheme.colorScheme.surface)) {
                     val (coverImage, taColumn, buttons) = createRefs()